refactor(device): change POLL_INTERVAL
Make POLL_INTERVAL static and no longer user definable. This ensures
that the App follows the Wallbox API guidelines. Current interval is set
to 61s to ensure that the API requests are distrubted as much as
possible.

Todo: ensure that UI settings for POLL_INTERVAL are removed.

Potential fix for #47.